add typing and move converters.ts
